Separate extract ddd concepts tests to clarify where we put specific concepts tests
